requirements.txt
langchain
ollama
openai
pandas
pgvector
psycopg2-binary
python-dotenv
quickumls
retriv
tiktoken